Added a separate global file for each world, They are named __ + the world name
Added in the ability to use absolute records, all sessions with owner tags are in absolute mode
Added in a timeout for log files so they can be closed after a while when they are no longer needed. Configurable through SpyPlugin.getSettings().logTimeout
Added updateSource() to SpreadTracker so that after a placeholder cause has been resolved, it can be applied to all future spreads.

Fixed PriorityExecutor not actually executing anything
Fixed a bug in CauseFinder where it was using != instead of ==
Fixed a bug in SpreadTracker where the delayed removals didnt check to see if there was anything still in the list after removing one

Removed the cause finding for tnt detonation, it can only possibly find the cause if the tnt doesnt move from its location, and it is causing massive issues if you detonate lots of tnt

Changed LogFileRegistry's unloadLogFile methods to call close, then remove if they are unloaded to make it more in line with what getLogFile does

Known bug in PriorityExecutor: All tasks get assigned to thread 0,