Bug Fixes:
Fixed NPE in InspectBlockTask when session failed to load
Fixed major issue in AbstractHoleIndex where it would use the normal spacelocator to consume the space, but this causes writes to happen corrupting the log or creating unallocated space
Fixed an issue in HoleIndex(tracdata) where the padding was being included in the size inside updateSize

LoadSession returns all the records that succeeded to load even if an error occurs.
Errors for failed session loads will only appear once in the console to help stop error spamming.
Added Horses, wolves, cats, sheep, and mooshrooms to breeding tracker
NOTE: Horses do not work with any trackers at the moment due to the null entity type problem